refactor: Extract config and extract check functions from index
