--- packages/repo-doctor/repositories.json
@@ -1,21 +1,0 @@
-[
-  {
-    "slug": "cozy/cozy-drive"
-  },
-  {
-    "slug": "cozy/cozy-banks"
-  },
-  {
-    "slug": "cozy/cozy-store"
-  },
-  {
-    "slug": "cozy/cozy-settings"
-  },
-  {
-    "slug": "cozy/cozy-home"
-  },
-  {
-     "slug": "cozy/cozy-contacts"
-  }
-]
-
